Divided the ListView operations as ViewOperations, and applied inheretenceto ListView and DetailedView. Moved rename item validation to BaseItem class. Removed not required imports in some classes. Implemented the create file operations and added the F7 hotkey and "New file - F7" button, to call the new create file method. Still missing the tests.
