+++ views/menus_toolbars/settings.py
@@ -1,0 +1,6 @@
+from PyQt4 import QtCore
+
+class Settings(QtCore.QSettings):
+    def __init__(self):
+        """Initializes the Settings class"""
+        super(Settings, self).__init__()
