+++ views/view_styles/__init__.py
@@ -1,3 +1,4 @@
 from list_view import ListView
 from icons_view import IconsView
 from detailed_view import DetailedView
+from view_operations import ViewOperations
