+++ views/view_styles/detailed_view.py
@@ -1,11 +1,13 @@
-
 from PyQt4 import QtGui
 from PyQt4 import QtCore
 
-from views.view_styles.list_view import ListView
+from view_operations import ViewOperations
 
-class DetailedView(ListView):
+class DetailedView(ViewOperations):
+	def __init__(self, current_path):
+		"""Sets the current path items view as list
 
-	def __init__(self):
-		"""Sets the currentPath and defines the panels"""
-		super(DetailedView, self).__init__(currentPath)
+		Params:
+		- current_path: receives the path to be set as current e.g. "C:\"
+		"""
+		super(DetailedView, self).__init__(current_path)
